+++ netpyne/specs/simConfig.py
@@ -44,8 +44,9 @@ class SimConfig (object):
         self.verbose = False  # show detailed messages
 
         # Recording
-        self.recordCells = []  # what cells to record from (eg. 'all', 5, or 'PYR')
+        self.recordCells = []  # what cells to record traces from (eg. 'all', 5, or 'PYR')
         self.recordTraces = {}  # Dict of traces to record
+        self.recordCellsSpikes = -1  # cells to record spike times from (-1 to record from all)
         self.recordStim = False  # record spikes of cell stims
         self.recordLFP = [] # list of 3D locations to record LFP from
         self.saveLFPCells = False  # Store LFP generate individually by each cell